Make DiffGraphProtoSerializer public.
